requirements.txt
flask==2.3.3
openai==2.15.0
python-dotenv==1.0.1
pypdf==4.0.1

